+++ connector-packager/connector_packager/package.py
@@ -1,5 +1,7 @@
-import os
+import argparse
 import logging
+import os
+import sys
 
 from argparse import ArgumentParser
 
@@ -14,16 +16,23 @@ from .version import __version__
 PACKAGED_EXTENSION = ".taco"
 
 
+class UniqueActionStore(argparse.Action):
+    def __call__(self, parser, namespace, values, option_string):
+        if getattr(namespace, self.dest, self.default) is not self.default:
+            parser.error(option_string + " appears > 1 times.")
+        setattr(namespace, self.dest, values)
+
+
 def create_arg_parser() -> ArgumentParser:
     parser = ArgumentParser(description="Tableau Connector Packaging Tool: package and sign connector files into a single Tableau Connector (" + PACKAGED_EXTENSION + ") file.")  # noqa: E501
     parser.add_argument('input_dir', help='path to directory of connector files to package and sign')
     parser.add_argument('-v', '--verbose', dest='verbose', action='store_true', help='verbose output', required=False)
-    parser.add_argument('-l', '--log', dest='log_path', help='path to directory for logging output',
+    parser.add_argument('-l', '--log', dest='log_path', help='Path to directory for log output. Logs are saved in the file packaging_logs.txt. Default is the current working directory.',
                         default=os.getcwd())
     parser.add_argument('--validate-only', dest='validate_only', action='store_true',
                         help='runs package validation steps only', required=False)
     parser.add_argument('-d', '--dest', dest='dest', help='destination folder for packaged connector',
-                        default='packaged-connector')
+                        default='packaged-connector', action=UniqueActionStore)
     parser.add_argument('--package-only', dest='package_only', action='store_true',
                         help='package a taco only, skip signing', required=False)
     parser.add_argument('-a', '--alias', dest='alias',
@@ -33,7 +42,7 @@ def create_arg_parser() -> ArgumentParser:
     return parser
 
 
-def init_logging(log_path: str, verbose: bool = False) -> logging.Logger:
+def init_logging(log_path: Path, verbose: bool = False) -> logging.Logger:
     # Create logger.
     logger = logging.getLogger('packager_logger')
     logger.setLevel(logging.DEBUG)
@@ -60,10 +69,25 @@ def init_logging(log_path: str, verbose: bool = False) -> logging.Logger:
     return logger
 
 
+def log_path_checker(path_to_logs: str) -> Path:
+    proper_path = Path(path_to_logs)
+    if proper_path.exists():
+        print("The log path {} exists".format(str(proper_path)))
+    else:
+        print("The specified log path does not exist - attempting to create {}".format(proper_path))
+        try:
+            proper_path.mkdir()
+            logging.info("Created {}".format(proper_path))
+        except Exception:
+            print("Unable to create log directory. Exiting.")
+            sys.exit(-1)
+    return proper_path / 'packaging_logs.txt'
+
+
 def main():
     parser = create_arg_parser()
     args = parser.parse_args()
-    log_file = args.log_path + '/packaging_log.txt'
+    log_file = log_path_checker(args.log_path)
     logger = init_logging(log_file, args.verbose)
 
     path_from_args = Path(args.input_dir)
@@ -79,7 +103,7 @@ def main():
     if files_to_package and validate_all_xml(files_to_package, path_from_args):
         logger.info("Validation succeeded.")
     else:
-        logger.info("Validation failed. Check " + log_file + " for more information.")
+        logger.info("Validation failed. Check " + str(log_file) + " for more information.")
         return
 
     # Double check that all files exist
@@ -98,7 +122,7 @@ def main():
     package_name = xmlparser.class_name + PACKAGED_EXTENSION
 
     if not jdk_create_jar(path_from_args, files_to_package, package_name, package_dest_path):
-        logger.info("Taco packaging failed. Check " + log_file + " for more information.")
+        logger.info("Taco packaging failed. Check " + str(log_file) + " for more information.")
         return
 
     if args.package_only:
@@ -109,11 +133,11 @@ def main():
     keystore_from_args = args.keystore
 
     if not validate_signing_input(package_dest_path, package_name, alias_from_args, keystore_from_args):
-        logger.debug("Signing input validation failed. check " + log_file + " for more information.")
+        logger.debug("Signing input validation failed. check " + str(log_file) + " for more information.")
         return
 
     if not jdk_sign_jar(package_dest_path, package_name, alias_from_args, keystore_from_args):
-        logger.info("Signing failed. check console output and " + log_file + " for more information.")
+        logger.info("Signing failed. check console output and " + str(log_file) + " for more information.")
         return